Feat: был переписан код по ТЗ

+++ commerce/shopping-cart/src/main/resources/schema.sql
@@ -8,6 +8,6 @@ CREATE TABLE IF NOT EXISTS cart_items (
     id BIGSERIAL PRIMARY KEY,
     cart_id BIGINT NOT NULL REFERENCES carts(id) ON DELETE CASCADE,
     product_id UUID NOT NULL,
-    quantity INT NOT NULL CHECK (quantity > 0),
+    quantity BIGINT NOT NULL CHECK (quantity > 0),
     UNIQUE (cart_id, product_id)
 );
